requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
argcomplete==3.2.3
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
cfn-flip==1.3.0
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.4
distro==1.9.0
dj-database-url==2.1.0
django-cors-headers==4.3.1
durationpy==0.6
exceptiongroup==1.2.0
frozenlist==1.4.1
gunicorn==21.2.0
h11==0.14.0
hjson==3.1.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
install==1.3.5
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.4
kappa==0.6.0
langchain==0.1.11
langchain-community==0.0.27
langchain-core==0.1.30
langchain-openai==0.0.8
langchain-text-splitters==0.0.1
langsmith==0.1.23
MarkupSafe==2.1.5
marshmallow==3.21.1
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.13.3
orjson==3.9.15
packaging==23.2
pgvector==0.2.5
placebo==0.9.0
psycopg2-binary==2.9.9
pydantic==2.6.3
pydantic_core==2.16.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
s3transfer==0.10.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.28
sqlparse==0.4.4
tenacity==8.2.3
text-unidecode==1.3
tiktoken==0.6.0
toml==0.10.2
tqdm==4.66.2
troposphere==4.7.0
typing-inspect==0.9.0
typing_extensions==4.10.0
urllib3==2.0.7
Werkzeug==3.0.1
whitenoise==6.6.0
yarl==1.9.4
